Fix compilation bugs

+++ src/com/oltpbenchmark/ThreadBench.java
@@ -124,7 +124,7 @@ public class ThreadBench implements Thread.UncaughtExceptionHandler {
             assert sample.startNs >= nextStartNs;
 
             // Collect all samples in the time window
-            ArrayList<Integer> latencies = new ArrayList<Integer>();
+            ArrayList<Long> latencies = new ArrayList<Long>();
             long endNs = nextStartNs + windowSizeSeconds * 1000000000L;
             while (sample != null && sample.startNs < endNs) {
 
@@ -145,7 +145,7 @@ public class ThreadBench implements Thread.UncaughtExceptionHandler {
             assert sample == null || endNs <= sample.startNs;
             nextStartNs = endNs;
 
-            int[] l = new int[latencies.size()];
+            long[] l = new long[latencies.size()];
             for (int i = 0; i < l.length; ++i) {
                 l[i] = latencies.get(i);
             }
@@ -507,7 +507,7 @@ public class ThreadBench implements Thread.UncaughtExceptionHandler {
             Collections.sort(samples);
 
             // Compute stats on all the latencies
-            int[] latencies = new int[samples.size()];
+            long[] latencies = new long[samples.size()];
             for (int i = 0; i < samples.size(); ++i) {
                 latencies[i] = samples.get(i).latencyUs;
             }
